backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML
openai>=1.10.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# HTTP
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.0

# Parsing
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0

# Optional: LangGraph (if you want to keep agent functionality)
# langgraph>=0.0.20
# langchain>=0.1.0
# langmem>=0.0.30
